🏷️ Gang Expansion and Item Update 🏷️
➕ Added rank multipliers
➕ Added multiple cooldowns: lootboxes, pay, gbank deposit, steal
➕ Added gang upgrades
➕ Added a gang vault to store and withdraw items
➕ Added a gank bank to deposit and withdraw money
➕ Added a gangkick command for an owner to kick unwanted members
➕ Added a gangmembers command to list all members of a gang
➕ Added so that all achivements that are preceeding the current awarded achivements are also given
➕ Added so that a gang also gets experience when a user works
➕ Added a gang experience system
➕ Added a -steal command
➕ Added two new leaderboards: rebirth and gang
➕ GangInvite now sends a user a DM notifying them of their invite
➕ Added 55 items
➕ Added gang tokens. Recieved when a gang levels up

🛠️ Fixed an issue with duplication of data in database
🛠️ Fixed an issue with gambling commands giving users too much money
🛠️ Fixed the bot crashing if a user types "-"
🛠️ Fixed an issue with lootcrates allowing you to buy even if you don't have the money.
🛠️ Fixed an issue with -jobs
🛠️ Fixed an issue with -gcreate showing names improperly
🛠️ Fixed some issues involving enoughMoney function
🛠️ Fixed an issue with all messages being removed if not in a command channel
🛠️ Fixed an issue with storing items
🛠️ Fixed an issue with achivements
🛠️ Fixed an issue with the -updaterules command
🛠️ Fixed an issue with law propositions storing

🗃️ Inventories can now be navigated using reaction buttons
🗃️ Changed the maximum amount of lootboxes
🗃️ Done some extra refactoring on the lootbox command
🗃️ Rebirths and suicide no longer reset gang, rebirths & achivements
🗃️ Stabilized most crates
🗃️ Changed the timeFormat function to show tenths if only seconds left
🗃️ Increased max work xp from 40 -> 50
🗃️ GangSettings now only toggles the setting
🗃️ Changed the timer for removing the inventory message from 20 to 45seconds
🗃️ Loot messages are now removed after 5 seconds
🗃️ Lootboxes can only be opened in the lootbox channels
🗃️ The help command now has multiple pages for certain categories
🗃️ Changed it so you can go below a balance of 0
🗃️ Changed the requirement of positives votes in a proposition

& m.m

🐞 0 Known

+++ events/guild/messageReactionAdd.js
@@ -1,17 +1,26 @@
 const updaterules = require('./../../loaders/updateRules.js');
 
-function checkVotes(reaction, client, disc){
-	const command = client.votes.get(reaction.message);
-	command[0].run(client, command[1]);
-	reaction.message.delete();
-	updaterules.update(client, disc, reaction.message);
-}
+let calculatePositivesNeeded = (negatives, members) => (5 * members * (negatives - 1) + 5 * ((n - 1) ** 2) + members ** 2) / (5 * members);
 
-//
+module.exports = (client, discord, reaction) => {
 
-module.exports = (client, disc, reaction) => {
-	if(reaction.message.channel.id == client.channelId.voting && reaction.emoji.name == '✅'
-		&& reaction.count >= Math.floor((reaction.message.guild.memberCount-client.botCount)*client.s.MAJORITY_RATE)+1){
-		checkVotes(reaction, client, disc);
-	}
+    if (reaction.message.channel.id == client.channelId.voting) {
+
+        const negativeCount = reaction.message.reactions.cache.get('❌').count;
+
+        if (negativeCount + Math.ceil(calculatePositivesNeeded(negativeCount, reaction.message.guild.memberCount - client.botCount)) > reaction.message.guild.memberCount - client.botCount) {
+            reaction.message.delete();
+        }
+
+        if (reaction.emoji.name == '✅' && reaction.count - 1 >= Math.ceil(calculatePositivesNeeded(negativeCount, reaction.message.guild.memberCount - client.botCount))) {
+            const command = client.votes.get(reaction.message);
+
+            if (command == undefined)
+                return;
+
+            command[0].run(client, command[1]);
+            reaction.message.delete();
+            updaterules.update(client, discord, reaction.message);
+        }
+    }
 }
